@@ -1,5 +1,5 @@
 {
-  "name": "starter-template",
+  "name": "quranicjibon",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -14,8 +14,8 @@
     "luxon": "^3.5.0",
     "npm": "^11.1.0",
     "prop-types": "^15.8.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.2.1",
+    "react-dom": "^19.2.1",
     "react-h5-audio-player": "^3.10.0-rc.1",
     "react-helmet-async": "^2.0.5",
     "react-icons": "^5.4.0",
